Working on controller
