--- controller/rss_feed_controller.py
@@ -1,34 +1,0 @@
-from classes.notification import Notification_Manager
-from model.rss_feed_model import RssFeedModel
-from classes.rss_feed import rss_feed_class
-from classes.validation import Validation
-from playsound import playsound
-from tkinter import messagebox
-import tkinter.font as tkFont
-from functools import partial
-from config import settings
-
-class rss_feed_controller:
-    def __init__(self):
-        # this will solved all the initials.
-        pass
-
-    def index(self):
-        # this will render the main page of this application.
-        pass
-
-    def store(self):
-        # this will store rss feeds from rss feed link.
-        pass
-
-    def show(self):
-        # this wil show the stored data from rss feed link to application.
-        pass
-
-    def delete(self):
-        # this will delete data from new tables.
-        pass
-
-    def update(self):
-        # this will move data from new tables to old tables.
-        pass
